Move dynamical logging from a separate module to core module
