+++ oap-server/pom.xml
@@ -48,7 +48,6 @@
         <module>server-tools</module>
         <module>server-fetcher-plugin</module>
         <module>server-health-checker</module>
-        <module>server-logging</module>
     </modules>
 
     <properties>
