+++ oap-server/server-bootstrap/pom.xml
@@ -245,11 +245,6 @@
             <artifactId>configuration-nacos</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.skywalking</groupId>
-            <artifactId>server-logging</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 
     <build>
